`NodeHTTPHandlerOptions`
