+++ packages/server/src/http/internals/HTTPHandlerOptions.ts
@@ -1,4 +1,3 @@
-import { Maybe } from '@trpc/server';
 import { OnErrorFunction } from '../../internals/OnErrorFunction';
 import {
   AnyRouter,
@@ -9,7 +8,6 @@ import {
 import { TRPCResponse } from '../../rpc';
 import { TRPCError } from '../../TRPCError';
 import { ResponseMeta } from '../ResponseMeta';
-import { HTTPRequest } from './types';
 
 type ResponseMetaFn<TRouter extends AnyRouter> = (opts: {
   data: TRPCResponse<unknown, inferRouterError<TRouter>>[];
@@ -23,7 +21,7 @@ type ResponseMetaFn<TRouter extends AnyRouter> = (opts: {
 }) => ResponseMeta;
 
 /**
- * Base interface for any HTTP/WSS handlers
+ * Base interface for any response handler
  */
 export interface BaseHandlerOptions<TRouter extends AnyRouter, TRequest> {
   onError?: OnErrorFunction<TRouter, TRequest>;
@@ -33,7 +31,10 @@ export interface BaseHandlerOptions<TRouter extends AnyRouter, TRequest> {
   router: TRouter;
 }
 
-export interface HTTPHandlerOptionsBase<TRouter extends AnyRouter, TRequest>
+/**
+ * Base interface for anything using HTTP
+ */
+export interface HTTPBaseHandlerOptions<TRouter extends AnyRouter, TRequest>
   extends BaseHandlerOptions<TRouter, TRequest> {
   /**
    * Add handler to be called before response is sent to the user
@@ -41,14 +42,4 @@ export interface HTTPHandlerOptionsBase<TRouter extends AnyRouter, TRequest>
    * @link https://trpc.io/docs/caching
    */
   responseMeta?: ResponseMetaFn<TRouter>;
-}
-
-export interface ResolveHTTPRequestOptions<
-  TRouter extends AnyRouter,
-  TRequest extends HTTPRequest,
-> extends HTTPHandlerOptionsBase<TRouter, TRequest> {
-  createContext: () => Promise<inferRouterContext<TRouter>>;
-  req: TRequest;
-  path: string;
-  error?: Maybe<TRPCError>;
 }
